refact : 특정 직관일지 조회 분리 #72
